requirements.txt
requests
nltk
python-dotenv
pandas
scipy
scikit-learn